frist setup liftPlannerDBDialog
